backend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.12
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==12.2.0
requests==2.32.3
